market-data-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
httpx == 0.28.1
requests == 2.32.5